Added fast path to parser for simple name lookup paths
Adds a "fast path" that bypasses the parser and constructs the path directly when it's just a series of named field lookups (e.g. 'foo.bar.baz').

In my tests this gives a ~2500x (!) speedup for parsing lots of simple paths of this type, which are common in our workflows.